requirements_test.txt
-r requirements.txt

coverage==4.4.1
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0
pytest-django>=4.1.0
pytest-runner>=5.2
pytest-env


# Additional test requirements go here
